try to add Word n-gram language model in order to generate Poems later, but currently is not working as expected.
Removed "using namespace std;" from every file where it was, and used "std::" instead.

Now input is just a line which contains words divided by spaces, but only last word is recognized for rhyming list of words. This change was made for future: for poem generator we need first sentence.